Resource path resolve, support for more properties
Added support for most of CSS properties.
Properties that are not dimensional will fallback to string from
now on.
Parsing of properties with multiple values will generate an array and
for consistency all dimensional values will be arrays.
Special consideration of margin and padding parsing was added.
Simplified the algorithm for file lookup and lowered the time/dependency
Support for box shadow, font-shadow.
Initial support for animations (WIP)